[ZENTORCH CORE] INT8 MoE Optimization (#641)
[ZENTORCH CORE] INT8 MoE Optimization
     - ZENAI-4686
     - GroupMatmul.cpp, Fused_Moe.cpp: Quantization of input to be pulled into FusedMoe.cpp.
     - _meta_registrations.py: Update the API with src_scales param.
     - test_group_matmul.py: Extend the testcase for the unique token-expert support.
     - zentorch_group_matmul.md, zentorch_fused_moe.md: Update design documents
     - Group the int8 rows with scale broadcast and run W13 with dynamic quant off,
       so ZenDNN does not re-quantize duplicated expert copies.
     - Fix single pass group_matmul with bf16 buffers filled with int8 values using direct_dynamic_quant API.

+++ src/cpu/python/zentorch/vllm/_int8_moe_patch.py
@@ -134,10 +134,10 @@ def _register_int8_moe_patches(mod) -> None:
             ), "CPUInt8Experts requires per-channel weight scales on the layer."
             E = self.w1_scale.shape[0]
             self._w13_scale = (
-                self.w1_scale.detach().to(torch.bfloat16).reshape(E, -1).contiguous()
+                self.w1_scale.detach().to(torch.float32).reshape(E, -1).contiguous()
             )
             self._w2_scale = (
-                self.w2_scale.detach().to(torch.bfloat16).reshape(E, -1).contiguous()
+                self.w2_scale.detach().to(torch.float32).reshape(E, -1).contiguous()
             )
             self._w13_bias = (
                 None
